+++ test/node/features/support/env.rb
@@ -12,3 +12,8 @@ Before('@skip_before_node_8') do |_scenario|
   node_version = ENV['NODE_VERSION'].to_i
   skip_this_scenario("Skipping scenario on Node #{node_version}") if node_version < 8
 end
+
+Before('@skip_before_node_16') do |_scenario|
+  node_version = ENV['NODE_VERSION'].to_i
+  skip_this_scenario("Skipping scenario on Node #{node_version}") if node_version < 16
+end